backend/requirements.txt
# REAL ENTERPRISE MATH TUTOR - ZERO COMPROMISES
# Strictly following assignment requirements - NO SQLite, ONLY Vector DB

# === CORE AGENT FRAMEWORKS (Assignment Required) ===
langchain>=0.2.0
langgraph>=0.0.60
langsmith>=0.1.20

# === VECTOR DATABASE ONLY (NO SQL ALLOWED) ===
# Qdrant - Professional Vector Database (as suggested in assignment)
qdrant-client>=1.7.0
# Weaviate - Alternative Vector Database (as suggested in assignment)  
weaviate-client>=3.25.0
# MongoDB Atlas Vector Search (cloud vector DB)
pymongo[srv]>=4.6.0
motor>=3.3.2

# === REAL EMBEDDINGS (NO SIMPLE IMPLEMENTATIONS) ===
sentence-transformers>=2.7.0
transformers>=4.38.0
torch>=2.2.0
datasets>=2.18.0

# === MODEL CONTEXT PROTOCOL (MANDATORY - ASSIGNMENT REQUIREMENT) ===
mcp>=1.0.0
httpx>=0.27.0

# === ADVANCED WEB SEARCH (Assignment Suggestions) ===
# Tavily - Professional Web Search API
tavily-python>=0.3.3
# Exa - Advanced Search Alternative
exa-py>=1.0.8
# Serper - Google Search API Alternative  
google-search-results>=2.4.2

# === DSPy FRAMEWORK (BONUS POINTS) ===
dspy-ai>=2.4.9
optuna>=3.5.0

# === AI GATEWAY & GUARDRAILS (ASSIGNMENT REQUIREMENT) ===
guardrails-ai>=0.4.5
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
pii-detector>=0.1.0

# === JEE BENCHMARKING (BONUS REQUIREMENT) ===
datasets>=2.17.0
huggingface-hub>=0.20.0
evaluate>=0.4.1
scikit-learn>=1.4.1
matplotlib>=3.8.3
seaborn>=0.13.2

# === FASTAPI & REACT REQUIREMENTS ===
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# === CORE DEPENDENCIES ===
numpy>=1.24.3
pandas>=2.0.3
python-dotenv>=1.0.0
aiofiles>=23.2.1
python-multipart>=0.0.6

# === TESTING & MONITORING ===
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0

# === NO SQLITE, NO SIMPLE DATABASES - VECTOR DB ONLY ===

# Monitoring & Logging
python-json-logger>=2.0.7
structlog>=23.2.0

# Environment Management
python-dotenv>=1.0.1

# HTTP Clients
aiohttp>=3.9.3
requests>=2.31.0

# Development Tools
pytest>=8.1.1
black>=24.2.0
isort>=5.13.2
mypy>=1.8.0

# Optional: OpenAI SDK (for better LLM performance)
openai>=1.12.0

# Optional: Anthropic SDK (for Claude)
anthropic>=0.21.3
